# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ai Botflow CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an owner and workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /auth/invitations/{token}:
    get:
      operationId: getInvitation
      tags: [auth]
      summary: Get a pending team invitation
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationInfo"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      tags: [auth]
      summary: Accept a team invitation
      parameters:
        - $ref: "#/components/parameters/Token"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationAcceptInput"
      responses:
        "200":
          description: Activated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    Token:
      name: token
      in: path
      required: true
      schema: { type: string, minLength: 20 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [owner, admin, agent] }
    Workspace:
      type: object
      required: [id, name, slug, plan]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [basic, starter, growth, enterprise] }
        walletBalance: { type: number }
...
    Session:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        user:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        workspace:
          anyOf:
            - $ref: "#/components/schemas/Workspace"
            - type: "null"
    RegisterInput:
      type: object
      required: [name, email, password, workspaceName]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string }
        password: { type: string, minLength: 8, maxLength: 128 }
        workspaceName: { type: string, minLength: 2, maxLength: 100 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string, minLength: 1, maxLength: 128 }
        workspaceSlug: { type: string, maxLength: 80 }
...
    InvitationInfo:
      type: object
      required: [workspaceName, email, role, expiresAt, needsPasswordSetup]
      properties:
        workspaceName: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, agent] }
        expiresAt: { type: string, format: date-time }
        needsPasswordSetup: { type: boolean }
    InvitationAcceptInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        password: { type: string, minLength: 8, maxLength: 128 }
...